refactor: use searchpairpos() and extmarks
Using vim.fn.searchpairpos and extmarks, we can
replace a large quantity of handcrafted
functionality with some maintained by vim and use
a more efficient way of creating highlights.

The benefit is shorter code which is easier to
modify and maintain and which exports labor onto
tools that are already built and used by the
editor itself.